+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,11 @@
 package racingcar;
 
+import racingcar.Controller.RacingGame;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
         RacingGame racingGame = new RacingGame();
-        racingGame.racingGame();
+        racingGame.startGame();
     }
 }
